Added zustand to the package. Refactored Settings Provider with Settings Store
